update warning window

+++ Client/include/AWidget.hh
@@ -5,7 +5,7 @@
 // Login   <brigno@epitech.net>
 //
 // Started on  Thu Oct 24 14:31:48 2013 brigno
-// Last update Wed Nov 20 17:33:10 2013 brigno
+// Last update Fri Nov 22 05:05:32 2013 brigno
 //
 
 #ifndef		__AWIDGET_HH__
@@ -29,20 +29,20 @@ public:
 protected:
   const sf::Event	&_event;
   std::string		_name;
-  sf::Vector2i		_posTopLeft;
-  sf::Vector2i		_focusTopLeft;
-  sf::Vector2i		_focusBotRight;
+  sf::Vector2f		_posTopLeft;
+  sf::Vector2f		_focusTopLeft;
+  sf::Vector2f		_focusBotRight;
   Type			_type;
 
 public:
-  AWidget(const sf::Event &event, const std::string &name, const sf::Vector2i &posTopLeft, const sf::Vector2i &focusTopLeft, const sf::Vector2i &focusBotRight, enum AWidget::Type type);
+  AWidget(const sf::Event &event, const std::string &name, const sf::Vector2f &posTopLeft, const sf::Vector2f &focusTopLeft, const sf::Vector2f &focusBotRight, enum AWidget::Type type);
   virtual ~AWidget();
   virtual MenuWindow::Status	onFocus(void) = 0;
   virtual void		stopFocus(void) = 0;
   virtual void		onHover(void) = 0;
   virtual void		stopHover(void) = 0;
   virtual void		draw(sf::RenderWindow &) = 0;
-  bool			isFocus(const sf::Vector2i &pos);
+  bool			isFocus(const sf::Vector2f &pos);
   const std::string	&getName(void) const {return (_name);};
   const Type		&getType(void) const {return (_type);};
 };
